Refactor main window code
